+++ Screens/LogIn.tsx
@@ -6,18 +6,15 @@ import { RootStackParams } from "../App";
 import Background from "../Components/Background";
 import HomeScreenButton from "../Components/Buttons/HomeScreenButton";
 import Logo from "../Components/Logo";
-import { useSound } from "../contexts/SoundContext";
+import CreateUser from "../Components/User/CreateUser";
+import UserExists from "../Components/User/UserExists";
 import { useTheme } from "../contexts/ThemeContext";
 import { User } from "../models/User";
-import CreateUserModal from "./CreateUserModal";
-import UserExistModal from "./UserExistModal";
 
 type Props = NativeStackScreenProps<RootStackParams, "LogIn">;
 
 const LogInScreen = ({ navigation }: Props) => {
   const { isDarkTheme, toggleTheme, themeColors } = useTheme();
-  const HomeScreenMusic = require("../assets/sounds/HomeScreenMusic.mp3");
-  const { playSound, toggleMuteMusic, toggleMuteButtonSound } = useSound();
   const [loginModalVisible, setLoginModalVisible] = useState(false);
   const [createUserModalVisible, setCreateUserModalVisible] = useState(false);
   const [isLoggedIn, setLoggedIn] = useState(false);
@@ -35,7 +32,7 @@ const LogInScreen = ({ navigation }: Props) => {
               setCreateUserModalVisible(!createUserModalVisible);
             }}
           >
-            <CreateUserModal
+            <CreateUser
               handleClose={() => {
                 setCreateUserModalVisible(false);
               }}
@@ -50,7 +47,7 @@ const LogInScreen = ({ navigation }: Props) => {
               setLoginModalVisible(!loginModalVisible);
             }}
           >
-            <UserExistModal
+            <UserExists
               handleClose={() => {
                 setLoginModalVisible(false);
               }}
